+++ tests/e2e/cases/dynamic_require_static_url_search_params/usp_entry_value_two.ts
@@ -1,0 +1,1 @@
+module.exports.label = "usp_entry_value_two";