+++ tests/e2e/cases/dynamic_require_static_url_search_params/usp_from_entries_3.ts
@@ -1,0 +1,1 @@
+module.exports.label = "usp_from_entries_3";